+++ ProyectoFinal/backend/src/main/java/com/uis/springboot/repository/CiudadRepository.java
@@ -4,7 +4,6 @@ import com.uis.springboot.model.Ciudad;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface CiudadRepository extends JpaRepository<Ciudad, Long> {
+public interface CiudadRepository extends JpaRepository<Ciudad, Integer> {
 }
