+++ ProyectoFinal/backend/src/main/java/com/uis/springboot/repository/TipoDocumentoRepository.java
@@ -4,8 +4,7 @@ import com.uis.springboot.model.TipoDocumento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface TipoDocumentoRepository extends JpaRepository<TipoDocumento, Long> {
-
+public interface TipoDocumentoRepository extends JpaRepository<TipoDocumento, Integer> {
 }
+
